[Windows/Desktop] Refactorization: fewer dependencies to Windows Headers definitions : CMake and CLion as Build/Development Alternative
<3tbs, Mummi and bilbo!!

Thomas ONE-MAN absolute <3!!

Mummi late dawn (from previous day till past 6:10AM Oct5) presence contribution!!

bilbo requesting to enter IT room (past 5:10 (AM)BRT) and laying on his tapestry: contribution!!

sodapoppin (past 5:10 (AM)BRT)  and friends contribution!!

iAssyrian chatting contribution!!

alfredotigolo, mandrakenk and Okbatgames chatting contribution!!

jitspoe, Endesga and Fearitself streaming contribution!!

tomrandall (past 19:10 BRT Oct5)  streaming Ludum contribution!!

vue_equalizer (fpga_guru) chatting contribution!!

EmptyGroceryBag (Joey) chatting after long time contribution!!

NorthSeaHero streaming and raid (2020-10-21 ~5:00) contribution!!

NorthSeaHero friends (ShyRyan,...) contribution!!

alfredotigolo chatting contribution!!

spoofhx (Spoofh, Fuel, Benjamin Kuhl from Köln) chatting contribution!!

Thiagovgamg chatting contribution!!

cmgriffing streaming contribution!!

Incarne streaming contribution!!

DonHo streaming contribution!!

lastmiles chatting contribution!!

phanxgames (Henri Price) streaming contribution!!

bytakos streaming contribution!!

Endesga C++ Development streaming  contribution!!

phaxgames (Henri Price) streaming contribution!!

OfficialAndyPyro streaming contribution!!

nextTab_de streaming contribution!!

nybblesio streaming contribution!!

Kebabskal streaming contribution!!

ThatDanishGuy streaming contribution!!

strawberryhacker streaming contribution!!

tomrandall streaming contribution!!

valdinei chatting contributtion after a long time!!

P4ndaExpress (Dunno) chatting contribution!!

SeppyTw and vue_equalizer (fpga_guru) talking contribution!!

CeeJaey streaming contribution (Mario 3D)!!

arn4vv and parrot_rl chatting contribution!!

lowluds chatting and interaction contribution!!

P4andaExpress (Dunno) and vue_equalizer (fpgaguru) chatting contribution!!

NorthSeaHero and FearItSelf streaming contribution!!

RetroMMO, GoPirateSoftware, lastmiles and togglebit streaming contribution!!

Ninja streaming contribution!!

GexYT streaming contribution!!

Beuca streaming contribution!!

Mads Kristensen@Visual Studio streaming contribution!!

JessicaMak, BendixBoy (& Friend) and AnderssonKev streaming contribution!!

strager and TheCherno streaming contribution!!

eren_kaynar7 chatting contribution!!

vue_equalizer (fpgaguru) chatting and ca2 development testing contribution!!

jhovgaard and tek256 streaming contribution!!

DonHo streaming contribution!!

nextTab_de streaming contribution!!

emergoentertainment streaming contribution!!

Ninja, Melchizedeck6809, Vimlark and strager streaming contribution!!

TimBeaudet and RareBytes (& friend) streaming contribution!!

unicorn_dev (& xmetrix) streaming contribution!!

Naysayer88 and friends contribution!!

ShemicalX streaming contribution!!

lelandkwong streaming contribution!!

oijtx streaming contribution!!

Ash_F0x and Myrkee streaming contribution!!

AdamCYounis and BarRaider streaming contribution!!

phanxgames and brianretrocorner streaming contribution!!

sodapoppin and friends contribution!!

jitspoe streaming contribution!!

Naysayer88 and friends contribution!!

lelandkwong streaming contribution!!

Goldbargames streaming contribution!!

Endesga (until 5:10(AM)BRT Oct5)  streaming contribution!!

jitspoe (till 18:10? BRT Oct5) and strager streaming contribution!!

Ash_F0x and JessicaMak streaming contribution!!

WTSRetro/SpiffyDane and Myrkee streaming contribution!!

Ninja and friends streaming contribution!!

erald_guri chatting contribution!!

lasteveq4 C++ interest chat contriubtion!!

cecilphillip and clarkio @"Microsoft Developer" streaming contribution!!

oijtx streaming contribution!!

NorthSeaHero streaming contribution!!

Trainwreckstv and friends streaming contribution!!

-- hi5 contribution...!!

from eddie machine running Windows 10 Pro remotely from bilbo machine running Windows 10 Pro!!

+++ acme/graphics/draw2d/color.h
@@ -225,10 +225,10 @@ public:
 
       double dComplement = 1.0 - dRate;
 
-      m_iR = m_iR * dComplement + color.m_iR * dRate;
-      m_iG = m_iG * dComplement + color.m_iG * dRate;
-      m_iB = m_iB * dComplement + color.m_iB * dRate;
-      m_iA = m_iA * dComplement + color.m_iA * dRate;
+      m_iR = (int) (m_iR * dComplement + color.m_iR * dRate);
+      m_iG = (int) (m_iG * dComplement + color.m_iG * dRate);
+      m_iB = (int) (m_iB * dComplement + color.m_iB * dRate);
+      m_iA = (int) (m_iA * dComplement + color.m_iA * dRate);
 
    }
