+++ apex/parallelization/fork.h
@@ -496,7 +496,7 @@ public:
    void construct()
    {
 
-      m_uThreadAffinityMask = (::uptr) translate_processor_affinity(m_iOrder);
+      m_uThreadAffinityMask = (::uptr) translate_processor_affinity((int) m_iOrder);
 
    }
